moontower-api/requirements.txt
requests>=2.31
pandas>=2.0
python-dotenv>=1.0
nbformat>=5.9
jupyter>=1.0
